+++ src/oneTrust/endpoints/index.ts
@@ -1,0 +1,3 @@
+export * from './getListOfOneTrustAssessments';
+export * from './getOneTrustAssessment';
+export * from './getOneTrustRisk';
