+++ src/oneTrust/index.ts
@@ -1,6 +1,3 @@
 export * from './createOneTrustGotInstance';
-export * from './getOneTrustAssessment';
-export * from './writeOneTrustAssessment';
-export * from './parseCliSyncOtArguments';
-export * from './getListOfOneTrustAssessments';
-export * from './getOneTrustRisk';
+export * from './helpers';
+export * from './endpoints';
